workflow-service/requirements.txt
pydantic-ai>=0.0.12
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.3.0
youtube-transcript-api>=0.6.1
pydantic>=2.5.0
python-dotenv>=1.0.0


